requirements.txt
celery==5.4.0
kombu==5.4.2
loguru==0.7.2
python-dotenv==1.0.1
pydantic==1.10.0
redis==5.0.8
beautifulsoup4==4.12.3
selenium==4.25.0
fastapi==0.115.0
requests==2.32.3
uvicorn==0.30.6
gevent==22.8.0
duckdb==1.1.1
pytz==2024.2
